Add files via upload
extra upload of data tables for mouse embryonic stem cell screens with sgVBC library haploid and diploid day 6 and day18.
We noticed that the old tables did not contain annotations for the 2 experimental timepoints "day6, day18" so it seemed that every datapoint was duplicated. Now we uploaded the more detailed data for day6 and day18 including individual data for each replicate. Each screen was carried out in 9-plicate.